requirements.txt
Django==4.2
python-gitlab==3.14.0
django-tinymce==3.6.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6